% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxplotCellScore.R
\name{BoxplotCellScore}
\alias{BoxplotCellScore}
\title{Boxplot of the CellScore values across test samples}
\usage{
BoxplotCellScore(cellscore, cell.change)
}
\arguments{
\item{cellscore}{a data.frame of CellScore values as calculated
by CellScore().}

\item{cell.change}{a data frame containing three columns, one for the
start (donor) test and target cell type. Each row of the data
frame describes one transition from the start to a target cell type.}
}
\value{
Invisibly, it returns list of the CellScore values by groups
  (in the same order as on the plot)
}
\description{
This function will plot a boxplot of the CellScore values for each selected
transition (defined in the cell.change data frame). The function will only
plot the scores for the test samples of valid subtypes (as annotated by
cellscore$sub_cell_type1). Scores for the standards are not included. Note
that if a subtype is specified by two different transitions, the coresponding
scores will be plotted in both transitions.
}
\examples{
## Load the expression set for the standard cell types
library(Biobase)
library(hgu133plus2CellScore) # eset.std

## Locate the external data files in the CellScore package
rdata.path <- system.file("extdata", "eset48.RData", package = "CellScore")
tsvdata.path <- system.file("extdata", "cell_change_test.tsv",
                             package = "CellScore")

if (file.exists(rdata.path) && file.exists(tsvdata.path)) {

   ## Load the expression set with normalized expressions of 48 test samples
   load(rdata.path)

   ## Import the cell change info for the loaded test samples
   cell.change <- read.delim(file= tsvdata.path, sep="\t",
                             header=TRUE, stringsAsFactors=FALSE)

   ## Combine the standards and the test data
   eset <- combine(eset.std, eset48)

   ## Generate cosine similarity for the combined data
   ## NOTE: May take 1-2 minutes on the full eset object
   ## so we subset it for 4 cell types
   pdata <- pData(eset)
   sel.samples <- pdata$general_cell_type \%in\% c("ESC", "EC", "FIB", "KER")
   eset.sub <- eset[, sel.samples]
   cs <- CosineSimScore(eset.sub, cell.change, iqr.cutoff=0.1)

   ## Generate the on/off scores for the combined data
   individ.OnOff <- OnOff(eset.sub, cell.change, out.put="individual")

   ## Generate the CellScore values for all samples
   cellscore <- CellScore(eset.sub, cell.change, individ.OnOff$scores,
                          cs$cosine.samples)

   ## Make the boxplot of CellScore values
   BoxplotCellScore(cellscore, cell.change)
}
}
\seealso{
\code{\link[CellScore]{CellScore}} for details on CellScore
  calculation.
}
\keyword{boxplot}
\keyword{cellscore}
