% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCosineSimHeatmap.R
\name{PlotCosineSimHeatmap}
\alias{PlotCosineSimHeatmap}
\title{Plot heatmap of the cosine similarity score}
\usage{
PlotCosineSimHeatmap(
  data,
  desc = "xx",
  width = 20,
  height = 20,
  x = -30,
  y = 3
)
}
\arguments{
\item{data}{a data.frame of cosine similarity scores, as generated
by the function CosineSimScore().}

\item{desc}{a single character, with description for the file name.
Suggested are "general.groups", "subgroups", and "samples".}

\item{width}{the width of the output pdf, in inches.}

\item{height}{the height of the output pdf, in inches.}

\item{x}{the x-position of the heatmap legend. It may be necessary to change
the value to position the legend in a suitable place on the plot.}

\item{y}{the y-position of the heatmap legend. It may be necessary to change
the value to position the legend in a suitable place on the plot.}
}
\value{
This function will print a pdf of the cosine similarity scores
in the current working directory.
}
\description{
This function plots a triangular heatmap of the cosine similarity scores.
}
\examples{
## Load the expression set for the standard cell types
library(Biobase)
library(hgu133plus2CellScore) # eset.std

## Locate the external data files in the CellScore package
rdata.path <- system.file("extdata", "eset48.RData", package = "CellScore")
tsvdata.path <- system.file("extdata", "cell_change_test.tsv",
                            package = "CellScore")

if (file.exists(rdata.path) && file.exists(tsvdata.path)) {

   ## Load the expression set with normalized expressions of 48 test samples
   load(rdata.path)

   ## Import the cell change info for the loaded test samples
   cell.change <- read.delim(file= tsvdata.path, sep="\t",
                             header=TRUE, stringsAsFactors=FALSE)

   ## Combine the standards and the test data
   eset <- combine(eset.std, eset48)

   ## Generate cosine similarity for the combined data
   ## NOTE: May take 1-2 minutes on the full eset object,
   ## so we subset it for 4 cell types
   pdata <- pData(eset)
   sel.samples <- pdata$general_cell_type \%in\% c("ESC", "EC", "FIB", "KER", 
                 "ASC", "NPC", "MSC")
   eset.sub <- eset[, sel.samples]
   cs <- CosineSimScore(eset.sub, cell.change, iqr.cutoff=0.1)

   ## Generate pdf of cosine similarity heatmap in the working directory
   PlotCosineSimHeatmap(cs$cosine.general.groups, "general groups",
                        width=7, height=7, x=-3.5, y=1)
}
}
\seealso{
\code{\link[CellScore]{CosineSimScore}} for details on cosine
  similarity calculation.
}
\keyword{cosine}
\keyword{heatmap}
\keyword{score,}
\keyword{similarity}
